requirements.txt
# HR Assistant — Dependencies
 
# PDF extraction
pdfplumber>=0.11.0
 
# Embeddings (local, gratuit)
sentence-transformers>=3.0.0
 
# Vector store
chromadb>=0.5.0
 
# LLM providers (installer selon le provider choisi)
openai>=1.40.0
anthropic>=0.34.0
mistralai>=1.0.0
groq>=0.9.0
 
# Web UI
streamlit>=1.38.0
 
# Env variables
python-dotenv>=1.0.0
 